mod-disco: dashboard and clean up house
