+++ mod-disco/service/flu/lib/modules/main_module.dart
@@ -1,24 +1,11 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:mod_disco/core/core.dart';
-import 'package:mod_disco/core/shared_repositories/mocks/mock_org_repository.dart';
-import 'package:mod_disco/core/shared_repositories/mocks/mock_support_role_answer_repository.dart';
-import 'package:mod_disco/core/shared_repositories/mocks/mock_support_role_repository.dart';
-import 'package:mod_disco/core/shared_repositories/mocks/mock_user_need_answer_repository.dart';
-import 'package:mod_disco/core/shared_repositories/mocks/mock_user_need_repository.dart';
+import 'package:mod_disco/modules/dashboard/views/org_master_detail_view.dart';
 import 'package:mod_disco/modules/new_orgs/views/org_view.dart';
-import 'package:mod_disco/modules/org_manager/orgs/views/org_master_detail_view.dart';
 import 'package:mod_disco/modules/projects/views/proj_view.dart';
-import 'package:mod_disco/modules/support_roles/services/support_role_answer_service.dart';
 import 'package:mod_disco/modules/survey_project/views/support_role_view.dart';
 import 'package:mod_disco/modules/survey_project/views/survey_project_view.dart';
-import 'package:mod_disco/modules/user_needs/services/user_need_answer_service.dart';
-import 'orgs/service/orgs_service.dart';
-import 'orgs/views/org_view.dart';
-import 'user_needs/services/user_need_service.dart';
 import 'splash/views/splash_view.dart';
-import 'support_roles/services/support_role_service.dart';
-import 'support_roles/views/support_role_view.dart';
-import 'user_needs/views/user_need_view.dart';
 import 'userinfo/views/userinfo_view.dart';
 
 class MainAppModule extends ChildModule {
@@ -44,18 +31,18 @@ class MainAppModule extends ChildModule {
   List<Bind> get binds => [
         Bind((i) => Paths(baseRoute)),
         Bind((i) => EnvConfig(url, urlNative)),
-        Bind((i) => OrgsService(repository: MockOrgRepository())),
-        // TODO Replace this later with OrgRepository
-        Bind((i) => UserNeedService(repository: MockUserNeedRepository())),
-        // TODO Replace this later with UserNeedRepository
-        Bind((i) =>
-            UserNeedAnswerService(repository: MockUserNeedAnswerRepository())),
-        // TODO Replace this later with UserNeedAnswerRepository
-        Bind(
-            (i) => SupportRoleService(repository: MockSupportRoleRepository())),
-        // TODO Replace this later with SupportRoleRepository
-        Bind((i) => SupportRoleAnswerService(
-            repository: MockSupportRoleAnswerRepository())),
+        // Bind((i) => OrgsService(repository: MockOrgRepository())),
+        // // TODO Replace this later with OrgRepository
+        // Bind((i) => UserNeedService(repository: MockUserNeedRepository())),
+        // // TODO Replace this later with UserNeedRepository
+        // Bind((i) =>
+        //     UserNeedAnswerService(repository: MockUserNeedAnswerRepository())),
+        // // TODO Replace this later with UserNeedAnswerRepository
+        // Bind(
+        //     (i) => SupportRoleService(repository: MockSupportRoleRepository())),
+        // // TODO Replace this later with SupportRoleRepository
+        // Bind((i) => SupportRoleAnswerService(
+        //     repository: MockSupportRoleAnswerRepository())),
         // TODO Replace this later with SupportRoleRepository
       ];
 
@@ -69,38 +56,39 @@ class MainAppModule extends ChildModule {
           "/projects",
           child: (_, args) => ProjectView(),
         ),
-        ModularRouter("/projects/:orgId/:id",
-            child: (_, args) => ProjectView(
-                  orgId: args.params['orgId'] ?? '',
-                  id: args.params['id'] ?? '',
-                )),
-        ModularRouter("/survey/projects/",
-            child: (_, args) => SurveyProjectView(
-                  project: args.data,
-                )),
-        ModularRouter("/support_roles/projects/",
-            child: (_, args) => SurveySupportRoleView(
-                  project: args.data['project'],
-                  surveyUserRequest: args.data['surveyUserRequest'],
-                  accountId: args.data['accountId'],
-                  surveyProjectList: args.data['surveyProjectList'],
-                )),
-        ModularRouter("/myneeds/projects/:id",
-            child: (_, args) => UserNeedsView(
-                  orgID: args.params['id'],
-                )),
-        ModularRouter("/supportRoles/projects/:id",
-            child: (_, args) => SupportRoleView(
-                  orgId: args.params['id'],
-                )),
+        ModularRouter(
+          "/projects/:orgId/:id",
+          child: (_, args) => ProjectView(
+            orgId: args.params['orgId'] ?? '',
+            id: args.params['id'] ?? '',
+          ),
+        ),
+        ModularRouter(
+          "/survey/projects/",
+          child: (_, args) => SurveyProjectView(
+            project: args.data,
+          ),
+        ),
+        ModularRouter(
+          "/support_roles/projects/",
+          child: (_, args) => SurveySupportRoleView(
+            project: args.data['project'],
+            surveyUserRequest: args.data['surveyUserRequest'],
+            accountId: args.data['accountId'],
+            surveyProjectList: args.data['surveyProjectList'],
+          ),
+        ),
 
         /// Admin Dashboard Routes
         ModularRouter("/dashboard/orgs",
             child: (_, args) => OrgMasterDetailView()),
-        ModularRouter("/dashboard/orgs/:id",
-            child: (_, args) => OrgMasterDetailView(
-                  id: int.tryParse(args.params['id']) ?? -1,
-                )),
+        ModularRouter(
+          "/dashboard/orgs/:id",
+          child: (_, args) => OrgMasterDetailView(
+            id: args.params['id'] ?? '',
+            orgId: args.params['orgId'] ?? '',
+          ),
+        ),
       ];
 
   static Inject get to => Inject<MainAppModule>.of();
