--- mod-disco/service/flu/lib/modules/org_manager/NOTE.md
@@ -1,1 +1,0 @@
-# This Directory is Deprecated
