+++ mod-disco/service/flu/lib/modules/projects/view_model/project_detail_view_model.dart
@@ -15,7 +15,7 @@ class ProjectDetailViewModel extends BaseModel {
 
   bool get isLoading => _isLoading;
 
-  DiscoProject get projectDetails => _discoProject;
+  DiscoProject get selectedProjectDetails => _discoProject;
 
   Future<void> fetchProjectDetail() async {
     _setLoading(true);
