+++ mod-disco/service/flu/lib/modules/projects/views/proj_detail_view.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:intl/intl.dart';
-import 'package:mod_disco/modules/projects/view_model/project_detail_view_model.dart';
 import 'package:mod_disco/modules/projects/views/proj_video_widget.dart';
 import 'package:mod_disco/rpc/v2/mod_disco_models.pb.dart';
-import 'package:provider_architecture/_viewmodel_provider.dart';
 import 'package:sys_share_sys_account_service/sys_share_sys_account_service.dart';
 import './proj_header.dart';
 import '../../../core/core.dart';
