--- mod-disco/service/flu/lib/modules/support_roles/services/support_role_service.dart
@@ -1,21 +1,0 @@
-import '../data/support_role_model.dart';
-import 'package:mod_disco/core/shared_repositories/support_role_repository.dart';
-
-class SupportRoleService {
-  final SupportRoleRepository _repository;
-
-  SupportRoleService({repository}) : this._repository = repository;
-
-  SupportRole getSupportRoleById(String id) {
-    return this._repository.getById(id);
-  }
-
-  List<SupportRole> getSupportRolesByOrgId(String orgId) {
-    return this._repository.getByOrgId(orgId);
-  }
-
-  List<SupportRole> getAll() {
-    // TODO Pull data from go client
-    return this._repository.getAll();
-  }
-}
