--- mod-disco/service/flu/lib/modules/user_needs/NOTE.md
@@ -1,1 +1,0 @@
-# This Directory is Deprecated
